+++ src/pages/calendar.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import GnbNavLayout from '@src/layouts/GnbNavLayout';
-import MonthlyCalendar from '@src/features/calendar/components/MothlyCalendar';
+import MonthlyCalendar from '@src/features/calendar/components/MonthlyCalendar';
 
 const Calendar = () => {
   return (
